+++ module-domain/src/test/java/com/depromeet/service/memory/TimelineServiceTest.java
@@ -1,4 +1,4 @@
-package com.depromeet.service;
+package com.depromeet.service.memory;
 
 import static org.assertj.core.api.Assertions.assertThat;
