+++ module-presentation/src/test/java/com/depromeet/image/api/ImageControllerTest.java
@@ -12,9 +12,9 @@ import com.depromeet.config.mock.WithCustomMockMember;
 import com.depromeet.image.dto.request.ImageNameRequest;
 import com.depromeet.image.dto.response.ImageResponse;
 import com.depromeet.image.dto.response.ImageUploadResponse;
+import com.depromeet.image.dto.response.ImageUploadResponseDtoFixture;
+import com.depromeet.image.dto.response.MemoryImagesDtoFixture;
 import com.depromeet.image.facade.ImageFacade;
-import com.depromeet.image.fixture.dto.response.ImageUploadResponseDtoFixture;
-import com.depromeet.image.fixture.dto.response.MemoryImagesDtoFixture;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.HashMap;
 import java.util.List;
