+++ module-presentation/src/test/java/com/depromeet/image/dto/response/ImageUploadResponseDtoFixture.java
@@ -1,6 +1,5 @@
-package com.depromeet.image.fixture.dto.response;
+package com.depromeet.image.dto.response;
 
-import com.depromeet.image.dto.response.ImageUploadResponse;
 import java.util.ArrayList;
 import java.util.List;
