+++ module-presentation/src/test/java/com/depromeet/image/dto/response/MemoryImagesDtoFixture.java
@@ -1,6 +1,5 @@
-package com.depromeet.image.fixture.dto.response;
+package com.depromeet.image.dto.response;
 
-import com.depromeet.image.dto.response.ImageResponse;
 import java.util.ArrayList;
 import java.util.List;
